+++ com.unity.render-pipelines.core/Runtime/Utilities/CoreUtils.cs
@@ -35,6 +35,7 @@ namespace UnityEngine.Rendering
         public const int editMenuPriority3 = 342;
         public const int assetCreateMenuPriority1 = 230;
         public const int assetCreateMenuPriority2 = 241;
+        public const int assetCreateMenuPriority3 = 300;
         public const int gameObjectMenuPriority = 10;
 
         static Cubemap m_BlackCubeTexture;