+++ com.unity.render-pipelines.universal/Runtime/UniversalAdditionalCameraData.cs
@@ -19,6 +19,7 @@ namespace UnityEngine.Rendering.Universal
         UsePipelineSettings,
     }
 
+    //[Obsolete("Renderer override is no longer used, renderers are referenced by index on the pipeline asset.")]
     [MovedFrom("UnityEngine.Rendering.LWRP")] public enum RendererOverrideOption
     {
         Custom,
@@ -58,9 +59,7 @@ namespace UnityEngine.Rendering.Universal
         [SerializeField]
         CameraOverrideOption m_RequiresOpaqueTextureOption = CameraOverrideOption.UsePipelineSettings;
 
-        [SerializeField] RendererOverrideOption m_RendererOverrideOption = RendererOverrideOption.UsePipelineSettings;
-        [SerializeField] ScriptableRendererData m_RendererData = null;
-        ScriptableRenderer m_Renderer = null;
+        [SerializeField] int m_RendererIndex = -1;
 
         [SerializeField] LayerMask m_VolumeLayerMask = 1; // "Default"
         [SerializeField] Transform m_VolumeTrigger = null;
@@ -136,16 +135,16 @@ namespace UnityEngine.Rendering.Universal
 
         public ScriptableRenderer scriptableRenderer
         {
-            get
-            {
-                if (m_RendererOverrideOption == RendererOverrideOption.UsePipelineSettings || m_RendererData == null)
-                    return UniversalRenderPipeline.asset.scriptableRenderer;
+            get => UniversalRenderPipeline.asset.GetRenderer(m_RendererIndex);
+        }
 
-                if (m_RendererData.isInvalidated || m_Renderer == null)
-                    m_Renderer = m_RendererData.InternalCreateRenderer();
-
-                return m_Renderer;
-            }
+        /// <summary>
+        /// Use this to set this Camera's current ScriptableRenderer to one listed on the Render Pipeline Asset. Takes an index that maps to the list on the Render Pipeline Asset.
+        /// </summary>
+        /// <param name="index">The index that maps to the RendererData list on the currently assigned Render Pipeline Asset</param>
+        public void SetRenderer(int index)
+        {
+            m_RendererIndex = index;
         }
 
         public LayerMask volumeLayerMask